Custom diagrams are supported everywhere now
