+++ core/src/main/java/org/eclipse/epsilon/labs/playground/fn/ModelLoader.java
@@ -12,6 +12,8 @@ import org.eclipse.emf.ecore.xmi.impl.XMIResourceFactoryImpl;
 import org.eclipse.emf.emfatic.core.EmfaticResource;
 import org.eclipse.epsilon.emc.emf.InMemoryEmfModel;
 import org.eclipse.epsilon.emc.json.JsonModel;
+import org.eclipse.epsilon.eol.EolModule;
+import org.eclipse.epsilon.eol.models.Model;
 import org.eclipse.epsilon.flexmi.FlexmiResourceFactory;
 
 import jakarta.inject.Singleton;
@@ -26,10 +28,6 @@ public class ModelLoader {
 	}
 
     public InMemoryEmfModel getInMemoryFlexmiModel(String flexmi, String emfatic) throws Exception {
-        return getInMemoryFlexmiModel(flexmi, emfatic, false);
-    }
-
-    public InMemoryEmfModel getInMemoryFlexmiModel(String flexmi, String emfatic, boolean annotated) throws Exception {
         ResourceSet resourceSet = new ResourceSetImpl();
         EPackage ePackage = getEPackage(emfatic);
         resourceSet.getPackageRegistry().put(ePackage.getNsURI(), ePackage);
@@ -39,13 +37,19 @@ public class ModelLoader {
         Resource resource = resourceSet.createResource(URI.createFileURI("/flexmi.flexmi"));
         resource.load(new ByteArrayInputStream(flexmi.getBytes()), null);
 
-        InMemoryEmfModel model = annotated ?
-                new AnnotatedInMemoryEmfModel(resource) :
-                new InMemoryEmfModel(resource);
+        InMemoryEmfModel model = isAnnotated(emfatic) ? new AnnotatableInMemoryEmfModel(resource) : new InMemoryEmfModel(resource);
         model.setName("M");
         return model;
     }
-    
+
+    protected boolean isAnnotated(String emfatic) throws Exception {
+        Model emfaticModel = getInMemoryEmfaticModel(emfatic);
+        EolModule checker = new EolModule();
+        checker.parse("return EClass.all.exists(c|c.eAnnotations.exists(a|a.source = 'node' or a.source = 'edge'));");
+        checker.getContext().getModelRepository().addModel(emfaticModel);
+        return (Boolean) checker.execute();
+    }
+
     public InMemoryEmfModel getInMemoryXmiModel(String xmi, String emfatic) throws Exception {
         ResourceSet resourceSet = new ResourceSetImpl();
         EPackage ePackage = getEPackage(emfatic);
@@ -54,7 +58,7 @@ public class ModelLoader {
         Resource resource = resourceSet.createResource(URI.createURI("xmi.xmi"));
         resource.load(new ByteArrayInputStream(xmi.getBytes()), null);
 
-        InMemoryEmfModel model = new InMemoryEmfModel(resource);
+        InMemoryEmfModel model = isAnnotated(emfatic) ? new AnnotatableInMemoryEmfModel(resource) : new InMemoryEmfModel(resource);
         model.setName("M");
         return model;
     }
@@ -66,8 +70,8 @@ public class ModelLoader {
 		resourceSet.getResourceFactoryRegistry().getExtensionToFactoryMap().put("*", new XMIResourceFactoryImpl());
 		Resource resource = resourceSet.createResource(URI.createURI("xmi.xmi"));
 
-		InMemoryEmfModel model = new InMemoryEmfModel(resource);
-		model.setName("M");
+        InMemoryEmfModel model = isAnnotated(emfatic) ? new AnnotatableInMemoryEmfModel(resource) : new InMemoryEmfModel(resource);
+        model.setName("M");
 		return model;
 	}
